Upgraded to new RingBuffer

+++ contracts/optimistic-ethereum/libraries/codec/Lib_OVMCodec.sol
@@ -83,7 +83,7 @@ library Lib_OVMCodec {
     struct QueueElement {
         bytes32 queueRoot;
         uint40 timestamp;
-        uint32 blockNumber;
+        uint40 blockNumber;
     }
 
     struct EOATransaction {
